requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptocode==0.1
cryptography==39.0.0
Flask==2.2.2
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.4
mypy-extensions==0.4.3
openai==0.26.0
pathspec==0.10.3
platformdirs==2.6.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.17
python-decouple==3.7
requests==2.28.1
rsa==4.9
SQLAlchemy==1.4.46
tomli==2.0.1
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
